Added generate external field
