--- yambopy/nl/generate_efield.py
@@ -1,5 +1,0 @@
-import sys
-import numpy as np
-import scipy as special
-import math  
-
